Count up in Both.
This is clunkier, but as @rachelmyers pointed out, it avoids the need
to reason on `increment` case-by-case.

For the same reason, it avoids the possibility of maintenance changing
`increment` in such a way that the selection of `first`/`second` falls
out of sync with the pattern being matched on; thus, it’s more
resilient to change.