Remove BranchInfo for If Syntaxes
Although this allowed us to report a list of summary statements for
each if / else-if condition, we don’t append parent contexts to
Patches. Because we lose that parent context, it makes replacement
summary statements devoid of context. Taking a step back, @joshvera and
I decided to leave if summary statements as they are now, but revisit
this when we explore rollups. Although diff summary statements do not
include else-if clauses now in Syntax If accurately. So we have the
info, but as it stands, a long list of summary statements for each
else-if clause we determined to be less helpful, but will set us up for
success when we add rollups.