Only remove indexed nodes for  import statements when the indexed node contains a child indexed node and so on.
This is to make sure for these cases, when we have an indexed node
containing a single indexed node and so on, that we’re capturing the
range of children Leaf or other Syntax other than Indexed nodes. This
is because of an issue that occurs when we construct SyntaxDiff for
import syntaxes specifically for nested indexed statements.