Add -A4m and -n2m GC flags
Sets the allocation area size to 4Mb and divides the allocation areas into chunks of 2mb. During execution, when a process exhausts its current chunk, it is given another chunk from the pool until the pool is exhausted, at which point a collection is triggered. This allows the processors to make better use of available allocation area, even when cores allocate at different rates.