Comment fixes to ensure `stack haddock` works.
There were a couple of typeclasses that were bamboozling the Haddock
parser. I was able to fix `subalgorithmFor`, but the vaguaries of
per-argument comments in type families appear to have escaped me—if
you can figure out how to get them picked up in the definition of
AssignmentParser, shout. I tried it all the ways I could think of and
came up empty.

But now we have nice Haddocks, which are intensely helpful!