Ensure gRPC ImportGraphs don't duplicate edges and vertices.
The previous code for generating an adjacency list from an algebraic
graph incorrectly assumed that each distinct vertex in the graph would
be `traverse`d once and only once. This is not true, as a vertex that
is connected to N other vertices will appear in N separate `Connect`
instances within the graph, and was leading to extreme redundancies in
graphs served over gRPC.

To work around this, as each vertex is encountered and tagged with a
unique value, the vertex-tag pair is stored in a hashmap. Subsequent
vertices are looked up in the environment; if there is a matching
hash, the original tag for that vertex is used.