Wait and block on the async parsing thread
If we timeout an async parsing thread and return, but the parsing thread is still running, we can force a race condition in which we delete the parser before it's completed on the async thread. This patch blocks on the parsing thread until it is complete before we clean up the parser. But this might have unintended consequences for the downstream consumer (i.e. github/github).

Co-Authored-By: Max Brunsfeld <maxbrunsfeld@gmail.com>