Regenerate data types from hydro.
Now that twirp-haskell generates significantly better code, we should
use it for our data types. This overcomes the significant problems
associated with JSON instances for protobuf objects, since now the
FromJSON and ToJSON instances are defined in terms of the JSONPB
instances, which brings them in line with the standard.

This is a breaking change, but we don't have any consumers of this API
yet, so @tclem gave me the all clear.