Disable optimizations for Util and REPL to speed up -O1 builds.
Turns out you can disable optimizations per-module. This is a good
decision, as on my laptop GHC takes an appalling amount of time to
finish building at `-O1`. Let's see if this is reflected in the time
the builders take. (This should have no performance implications for
CLI use of `semantic`, as only SpecHelpers and REPL import Util.)