Remove Error effect from tagging internals.
This code was based on code I wrote for the reprinter, which had
different correctness constraints than does the tagging system:
specifically, the tagging code does not fail, as there is a
recoverable result even if we get pathological input. As such, there's
no need to wrap this in an Either that is always Right.