Institute freeze file to lock dep versions for release builds.
This file, generated with `cabal new-freeze`, is used by the Docker
image to build semantic with a reproducible set of dependencies.

The Travis builders do not use this freeze file, since we want to make
sure that our PVP compliance doesn't suffer, and that anyone can just
pull the project and build it.