Add test for Semantic.Git.lsTree.
I have a proposed speedup for lsTree, but since this resides on a
critical path for downstream systems, I want to make sure that I don't
break this while I'm at it.

I really should rewrite this test to use tasty-hspec, but it'd be a
long slog, and I have better things to do.