Remove Data.Project.readFile and :fire: error handling in Graph.
Because `Data.Project.readFile` could throw an `FileNotFound` error
when looking up `Blob` values in a project for a given `File`, the
code that kicks off graphing required an error constraint. Turns out
we don't even need it: if we have a project, we can access its blobs
directly, since `Blob`s contain `File`s. This allows us to obviate
`Project.readFile` in the first place, which allows us to entirely
sidestep the need to handle errors.