Correctly account of the presence of autogen files in build/noopt.
If you have `optimizations: False` in your cabal.project.local (such
as when you're in a tight build-run-debug cycle for some executable or
test suite), the autogenerated cabal_macros and Paths_ files might not
be present in the `build/` folder; if they aren't, we should look in
`build/noopt`.

This patch appears to address an issue I saw where the cabal macros
were being redefined, but I'm not entirely sure why. :/