Silence REPL macro-redefined warnings from the C compiler.
In certain cases, use of the REPL can land us in a state where there
are multiple `cabal_macros.h` files being included, leading to
duplicate macro definitions. This does not factor into correctness,
as these duplicates are, well, duplicate definitions, but the warning
is tedious, so this patch passes `-Wno-macro-redefined` to the C compiler.