Remove Markdown support.
Markdown was always the proverbial redheaded stepchild of this
project: it was the only project not to use a tree-sitter
parsing (depending instead on `cmark-gfm`), it doesn't really work in
a tagging context, and has no computational content.

Now that internal projects have divested from the code paths that
invoked this functionality, we can remove it. Because Markdown is a
markup language rather than a programming language, it will probably
not return. Any open-source users (though I don't believe any exist)
should switch to `pandoc`.