Ensure that the failParsingForTesting flag is always respected.
The `runAssignment` code path makes sure to check that the
`failParsingForTesting` flag is set and throwing an error if
so; this patch ensures that the check happens on every code path.

I chose not to change the `runAssignment` code path to use this new
executeParserAction function, since that whole spiel is going away.