Fix annoying warning re. inter-project dependencies.
Building `semantic-python` would result in a warning like this:

```
Warning: The package has an extraneous version range for a dependency on an
internal library: semantic-python ==0.0.0.0. This version range includes the
current package but isn't needed as the current package's library will always
be used.
```

Removing this indeed-extraneous bound makes it go away.