Dynamically change cache keys based on cabal files' hash.
The problem with our caching approach up until now is that we've been
hitting Actions caches too often. As we `cabal update` and `cabal
configure`, our build plan strays farther and farther from what's
present in the cache, resulting in more and more cache missed from
`cabal`. The fix here is to hash a file that's changed reasonably
often and to use that to key the caches, falling back on the most
recently used cache present on a miss.